sam-app/requirements.txt
requests
aws-psycopg2
redis
cfnresponse
